+++ neuralqx/experimental/driver/__init__.py
@@ -16,7 +16,6 @@
 Experimental VMC drivers in neuraLQX.
 """
 
-from .mvmc import MultiStateVMC
 from .stmh_multi_vmc import SingleTrunkMultiHeadVMC
 
-__all__ = ["MultiStateVMC", "SingleTrunkMultiHeadVMC"]
+__all__ = ["SingleTrunkMultiHeadVMC"]